requirements.txt
asgiref==3.4.1
celery==5.2.1
Django==4.0
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
sqlparse==0.4.2
tzdata==2021.5